refactor: Produto
